pyproject.toml
[project]
name = "mcp-demo-project"
version = "0.1.0"
description = "A Demo Project to understand the working of MCP servers"
authors = [
    {name = "Pulkit Agarwal",email = "pulkit.agarwal@dianapps.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]

skip = [".venv"]

[tool.poetry]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
fastapi-mcp = "^0.3.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
pydantic = {extras = ["email"], version = "^2.10.0"}
pydantic-settings = "^2.6.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0"
email-validator = "^2.2.0"
greenlet = "^3.1.1"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
flake8 = "^7.3.0"
isort = "6.1.0"
pylint = "^3.3.9"
pylint-pydantic = "^0.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
exclude = '''
/(
  \.venv
)/
'''

[tool.isort]
profile = "black"
line_length = 88
